pyproject.toml
[project]
name = "signbot-backend"
version = "0.2.0"
description = "Secure web showcase + internal backend for Telegram signing bot"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "sqlmodel>=0.0.22",
  "pydantic-settings>=2.6.1",
  "python-multipart>=0.0.19",
  "jinja2>=3.1.4",
  "httpx>=0.27.2",
  "cryptography>=44.0.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.3"
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]
